This is proof of concept of significant `StreamReader` speedup. The main idea is avoiding memory copying and reallocations. The `collections.deque` is used instead of the `bytearray`, but unlike for `DataQueue`, no reading limitations are applied.

## Performance

My benchmarks: https://gist.github.com/homm/3b9c2909905f76668bf9

Test | old time | new time | speedup
-----|-----|-----|--------
Large chunk |
test_stream_feed | 0.0298 | 0.00001 | 2450x
test_stream_read_all | 0.2127 | 0.00005 | 3948x
test_stream_read_less | 0.2433 | 0.0506 | 4.8x
test_stream_read_any | 0.0608 | 0.00005 | 1128x
test_stream_read_exactly | 0.2005 | 0.0173 | 11.5x
Small chunks |
test_stream_feed | 0.0674 | 0.0094 | 7.1x
test_stream_read_all | 0.2633 | 0.0508 | 5.2x
test_stream_read_less | 0.2165 | 0.0556 | 3.9x
test_stream_read_any | 0.1022 | 0.0237 | 4.3x
test_stream_read_exactly | 0.1825 | 0.0371 | 4.9x

## Memory

As a result of removing relocations, the memory consumption and fragmentation is also greatly reduced. I've got following result for the benchmark:

* Without `test_stream_read_all` test.
  * Old: mostly 350-450MB, up to 550MB
  * New: 92MB
* With `test_stream_read_all` test.
  * Old: mostly 350-450MB, up to 550MB
  * New: 170MB

## Distinct

The main difference is how `read(BIG_LIMIT)` and `readany()` produce the result when the buffer is full: In the old implementation, the whole buffer is returned at once. In the new implementation, only the first chunk is returned.