@fafhrd91 2015-09-07
Sure! Please re-write readline(), do not touch tests, I'd like to test it on real system