@homm 2015-09-07
@fafhrd91 Done.

In my opinion `readline()` is totaly broken, because it checks limits after reading from _buffer and also it corrupts input stream if line size > limit, because it reads the undefined amount of data and lose it. But I think this is not the topic of current PR, and I've restored the old behavior.