@GMLudo 2015-09-07
Thank you a lot @homm, I'm pretty excited to test that with the [FrameworkBenchmarks](https://www.techempower.com/benchmarks/) suite to see the difference with the previous version of aiohttp.

Can I already test with this branch, or I must wait that you fix all the tests compliance first ?