@kxepal 2015-09-08
@homm Nice work! In additional to aiohttp test, it also breaks aiocouchdb ones:
```
Traceback (most recent call last):
  File "/home/kxepal/projects/aio-libs/aiocouchdb/aiocouchdb/tests/utils.py", line 36, in wrapper
    return testcase.loop.run_until_complete(future)
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/base_events.py", line 316, in run_until_complete
    return future.result()
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/futures.py", line 275, in result
    raise self._exception
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/tasks.py", line 238, in _step
    result = next(coro)
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/tasks.py", line 377, in wait_for
    return fut.result()
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/futures.py", line 275, in result
    raise self._exception
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/tasks.py", line 236, in _step
    result = coro.send(value)
  File "/home/kxepal/projects/aio-libs/aiocouchdb/aiocouchdb/v1/tests/test_document.py", line 232, in test_get_with_atts
    yield from result.release()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/multipart.py", line 187, in release
    yield from self.resp.release()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/client_reqrep.py", line 638, in release
    chunk = yield from content.readany()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/streams.py", line 391, in wrapper
    result = yield from func(self, *args, **kw)
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/streams.py", line 454, in readany
    return (yield from super().readany())
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/streams.py", line 199, in readany
    return self._read_nowait()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/streams.py", line 253, in _read_nowait
    self._buffer_size -= len(data)
TypeError: object of type 'int' has no len()
```

I tried to reproduce the problem with public instance, but here I also found that patch breaks multipart reader as well in other way:
```python
import asyncio
import aiohttp


@asyncio.coroutine
def go():
    resp = yield from aiohttp.request(
        'get',
        'https://kxepal.cloudant.com/aiohttp-496-bug/docid',
        headers={'Accept': 'multipart/related'},
        params={'attachments': 'true'})
    reader = aiohttp.MultipartReader.from_response(resp)
    while True:
        part = yield from reader.next()
        if part is None:
            break
        print((yield from part.read()))


if __name__ == '__main__':
    asyncio.get_event_loop().run_until_complete(go())
```

Expected:
```
bytearray(b'{"_id":"docid","_rev":"6-783d465b150e111b56447cd1d4e533a8","_attachments":{"test":{"content_type":"application/octet-stream","revpos":1,"digest":"md5-JS3VMT0Dk7FXE0dUvc9Vag==","length":15,"follows":true}}}')
bytearray(b'Hello, aiohttp!')
```

Actual:
```
Traceback (most recent call last):
  File "/home/kxepal/projects/aio-libs/aiocouchdb/test.py", line 21, in <module>
    asyncio.get_event_loop().run_until_complete(go())
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/base_events.py", line 316, in run_until_complete
    return future.result()
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/futures.py", line 275, in result
    raise self._exception
  File "/usr/local/opt/virtualenv/aiocouchdb/lib/python3.3/site-packages/asyncio/tasks.py", line 236, in _step
    result = coro.send(value)
  File "/home/kxepal/projects/aio-libs/aiocouchdb/test.py", line 14, in go
    part = yield from reader.next()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/multipart.py", line 178, in next
    item = yield from self.stream.next()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/multipart.py", line 460, in next
    self._last_part = yield from self.fetch_next_part()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/multipart.py", line 475, in fetch_next_part
    headers = yield from self._read_headers()
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/multipart.py", line 536, in _read_headers
    headers, *_ = parser.parse_headers(lines)
  File "/home/kxepal/projects/aio-libs/aiohttp/aiohttp/protocol.py", line 85, in parse_headers
    raise errors.InvalidHeader(name)
aiohttp.errors.InvalidHeader: 400, message='Invalid HTTP Header: 50E111B56447CD1D4E533A8","_ATTACHMENTS"'
```

`parser.parse_headers` receives the following `lines`:
```
['', 'content-type: application/json\r\n\r\n{"_id":"docid","_rev":"6-783d465b1', '50e111b56447cd1d4e533a8","_attachments":{"test":{"content_type":"application/octet-stream","revpos":1,"digest":"md5-JS3VMT0Dk7FXE0dUvc9Vag==","length":15,"follows":true}}}\r\n--c1782ac8c1a68da8b0bd4f8983b86fa3', '']
``` 
while it should stop right after '\r\n\r\n'. 

Thoughts? If you need any additional information, feel free to ask.