@homm 2015-09-08
@kxepal Thanks, I've fixed the error in `readline()`, so `MultipartReader` works again.

But I can't reproduce the error with `aiocouchdb`. I've ran tests from master and got **a lot** exceptions in `__del__` method of `client_reqrep.ClientResponse` (I have Python 3.4):

```
Traceback (most recent call last):
  File "/Users/Master/Code/async/aiohttp/aiohttp/client_reqrep.py", line 549, in __del__
    self._loop.call_exception_handler(context)
AttributeError: 'NoneType' object has no attribute 'call_exception_handler'
Exception ignored in: <bound method HttpResponse.__del__ of <ClientResponse() [200 None]>
{'CONTENT-TYPE': 'application/json'}
```

But except for these errors all tests are passed.

```
Ran 302 tests in 2.083s
OK (skipped=9)
```