@kxepal 2015-09-08
@homm Oh, my. It seems I didn't push the fixes. Thanks for notice!

The problem happens with `make check-couchdb` - simple `make check` runs tests against mocks and cannot || not able to trigger this issue. However, I found the error on my side. Ironically, it's within the [danger zone](https://github.com/aio-libs/aiocouchdb/blob/master/aiocouchdb/v1/document.py#L313) where I touch `_buffer` that you had changed (: s/extend/append/ solves the problem.

Sorry for false alarm (: