@homm 2015-09-08
@kxepal

> s/extend/append/ solves the problem.

Unfortunately no, because there is also `_buffer_size` and `total_bytes` properties. Why you can't use `feed_data` there?

I had the idea to create a wrapper around `collections.deque` with `bytearray` syntax (`extend` and `len`) for backward compatibility. But I don't know Is it really necessary for private property of `0.x` version library :—)