@kxepal 2015-09-08
@homm 
Actually, I don't know why. Strange decisions eventually happens (: Will investigate public API for better workaround.

As for idea to mimic deque to bytesarray, I think it doesn't worth any effort. In the end, field is private - there are no guarantees provided on them.